Milestone 3/requirements.txt
# Milestone 3 Requirements
playwright==1.40.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
python-dotenv==1.0.0
pytest==7.4.3
flask==3.0.0

